change tag renaming and moving shortcuts
tags are no longer directly renamed. A name can be applied, which is
appended (after a dash) to the default tag title. These names are retained
when moving tags, though the default title section is changed
accordingly so the tags will always reflect the shortcut key used to
interact with them.